Updated userModel, app-routing module, profile, error and home pages.
Updated userModel with the following, added a validation between the e-mail for it to be required, as well as maxValue for the password, so now it's 1-64 characters, not 5 to infinity.
Username stays the same.

Moved the ** and 404 paths, because they were not in the correct order.

Made the 404 page look 'better', as the styling was not the best.

Fixed the classes of the Description and Due Date on the home component, as well as making their own CSS Classes.

Removed the ImageURL from task type, since it wasn't really needed.